Replace sys.comp 1.1.0 with 9.0.0
