+++ src/referencePackages/src/microsoft.codeanalysis.workspaces.common/4.0.1/Customizations.props
@@ -6,11 +6,11 @@
 
     <!-- Upgrade System.Composition to a version that has netstandard2.0 -->
     <ProjectReference Remove="../../system.composition/1.0.31/System.Composition.1.0.31.csproj" />
-    <ProjectReference Include="../../system.composition.attributedmodel/1.1.0/System.Composition.AttributedModel.1.1.0.csproj" />
-    <ProjectReference Include="../../system.composition.convention/1.1.0/System.Composition.Convention.1.1.0.csproj" />
-    <ProjectReference Include="../../system.composition.hosting/1.1.0/System.Composition.Hosting.1.1.0.csproj" />
-    <ProjectReference Include="../../system.composition.runtime/1.1.0/System.Composition.Runtime.1.1.0.csproj" />
-    <ProjectReference Include="../../system.composition.typedparts/1.1.0/System.Composition.TypedParts.1.1.0.csproj" />
+    <ProjectReference Include="../../system.composition.attributedmodel/9.0.0/System.Composition.AttributedModel.9.0.0.csproj" />
+    <ProjectReference Include="../../system.composition.convention/9.0.0/System.Composition.Convention.9.0.0.csproj" />
+    <ProjectReference Include="../../system.composition.hosting/9.0.0/System.Composition.Hosting.9.0.0.csproj" />
+    <ProjectReference Include="../../system.composition.runtime/9.0.0/System.Composition.Runtime.9.0.0.csproj" />
+    <ProjectReference Include="../../system.composition.typedparts/9.0.0/System.Composition.TypedParts.9.0.0.csproj" />
   </ItemGroup>
 
 </Project>
