--- src/referencePackages/src/system.composition.attributedmodel/1.1.0/System.Composition.AttributedModel.1.1.0.csproj
@@ -1,9 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>netstandard2.0</TargetFrameworks>
-    <PackageVersion>1.1.0</PackageVersion>
-    <AssemblyName>System.Composition.AttributedModel</AssemblyName>
-  </PropertyGroup>
-
-</Project>
