--- src/referencePackages/src/system.composition.convention/1.1.0/System.Composition.Convention.1.1.0.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>netstandard2.0</TargetFrameworks>
-    <PackageVersion>1.1.0</PackageVersion>
-    <AssemblyName>System.Composition.Convention</AssemblyName>
-  </PropertyGroup>
-
-  <ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.0'">
-    <ProjectReference Include="../../system.composition.attributedmodel/1.1.0/System.Composition.AttributedModel.1.1.0.csproj" />
-  </ItemGroup>
-
-</Project>
