--- src/referencePackages/src/system.composition.convention/1.1.0/system.composition.convention.nuspec
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<package xmlns="http://schemas.microsoft.com/packaging/2013/05/nuspec.xsd">
-  <metadata minClientVersion="2.12">
-    <id>System.Composition.Convention</id>
-    <version>1.1.0</version>
-    <title>System.Composition.Convention</title>
-    <authors>Microsoft</authors>
-    <owners>microsoft,dotnetframework</owners>
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <licenseUrl>https://github.com/dotnet/corefx/blob/master/LICENSE.TXT</licenseUrl>
-    <projectUrl>https://dot.net/</projectUrl>
-    <iconUrl>http://go.microsoft.com/fwlink/?LinkID=288859</iconUrl>
-    <description>Provides types that support using Managed Extensibility Framework with a convention-based configuration model.
-
-Commonly Used Types:
-System.Composition.Convention.ConventionBuilder
-System.Composition.Convention.ExportConventionBuilder
-System.Composition.Convention.ImportConventionBuilder
-System.Composition.Convention.PartConventionBuilder
-System.Composition.Convention.ParameterImportConventionBuilder
- 
-8321c729934c0f8be754953439b88e6e1c120c24 
-When using NuGet 3.x this package requires at least version 3.4.</description>
-    <releaseNotes>https://go.microsoft.com/fwlink/?LinkID=799421</releaseNotes>
-    <copyright>© Microsoft Corporation.  All rights reserved.</copyright>
-    <serviceable>true</serviceable>
-    <dependencies>
-      <group targetFramework=".NETStandard2.0">
-        <dependency id="System.Composition.AttributedModel" version="1.1.0" />
-      </group>
-    </dependencies>
-  </metadata>
-</package>
