--- src/referencePackages/src/system.composition.hosting/1.1.0/System.Composition.Hosting.1.1.0.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>netstandard2.0</TargetFrameworks>
-    <PackageVersion>1.1.0</PackageVersion>
-    <AssemblyName>System.Composition.Hosting</AssemblyName>
-  </PropertyGroup>
-
-  <ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.0'">
-    <ProjectReference Include="../../system.composition.runtime/1.1.0/System.Composition.Runtime.1.1.0.csproj" />
-  </ItemGroup>
-
-</Project>
