--- src/referencePackages/src/system.composition.runtime/1.1.0/System.Composition.Runtime.1.1.0.csproj
@@ -1,9 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>netstandard2.0</TargetFrameworks>
-    <PackageVersion>1.1.0</PackageVersion>
-    <AssemblyName>System.Composition.Runtime</AssemblyName>
-  </PropertyGroup>
-
-</Project>
