--- src/referencePackages/src/system.composition.runtime/1.1.0/system.composition.runtime.nuspec
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<package xmlns="http://schemas.microsoft.com/packaging/2013/05/nuspec.xsd">
-  <metadata minClientVersion="2.12">
-    <id>System.Composition.Runtime</id>
-    <version>1.1.0</version>
-    <title>System.Composition.Runtime</title>
-    <authors>Microsoft</authors>
-    <owners>microsoft,dotnetframework</owners>
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <licenseUrl>https://github.com/dotnet/corefx/blob/master/LICENSE.TXT</licenseUrl>
-    <projectUrl>https://dot.net/</projectUrl>
-    <iconUrl>http://go.microsoft.com/fwlink/?LinkID=288859</iconUrl>
-    <description>Contains runtime components of the Managed Extensibility Framework.
-
-Commonly Used Types:
-System.Composition.CompositionContext
- 
-8321c729934c0f8be754953439b88e6e1c120c24 
-When using NuGet 3.x this package requires at least version 3.4.</description>
-    <releaseNotes>https://go.microsoft.com/fwlink/?LinkID=799421</releaseNotes>
-    <copyright>© Microsoft Corporation.  All rights reserved.</copyright>
-    <serviceable>true</serviceable>
-    <dependencies>
-      <group targetFramework=".NETStandard2.0" />
-    </dependencies>
-  </metadata>
-</package>
