+++ src/referencePackages/src/system.composition.runtime/9.0.0/System.Composition.Runtime.9.0.0.csproj
@@ -1,0 +1,9 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFrameworks>net8.0;net9.0;netstandard2.0</TargetFrameworks>
+    <PackageVersion>9.0.0</PackageVersion>
+    <AssemblyName>System.Composition.Runtime</AssemblyName>
+  </PropertyGroup>
+
+</Project>
