--- src/referencePackages/src/system.composition.typedparts/1.1.0/System.Composition.TypedParts.1.1.0.csproj
@@ -1,15 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>netstandard2.0</TargetFrameworks>
-    <PackageVersion>1.1.0</PackageVersion>
-    <AssemblyName>System.Composition.TypedParts</AssemblyName>
-  </PropertyGroup>
-
-  <ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.0'">
-    <ProjectReference Include="../../system.composition.attributedmodel/1.1.0/System.Composition.AttributedModel.1.1.0.csproj" />
-    <ProjectReference Include="../../system.composition.hosting/1.1.0/System.Composition.Hosting.1.1.0.csproj" />
-    <ProjectReference Include="../../system.composition.runtime/1.1.0/System.Composition.Runtime.1.1.0.csproj" />
-  </ItemGroup>
-
-</Project>
