+++ src/referencePackages/src/system.composition.typedparts/9.0.0/System.Composition.TypedParts.9.0.0.csproj
@@ -1,0 +1,27 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFrameworks>net8.0;net9.0;netstandard2.0</TargetFrameworks>
+    <PackageVersion>9.0.0</PackageVersion>
+    <AssemblyName>System.Composition.TypedParts</AssemblyName>
+  </PropertyGroup>
+
+  <ItemGroup Condition="'$(TargetFramework)' == 'net8.0'">
+    <ProjectReference Include="../../system.composition.attributedmodel/9.0.0/System.Composition.AttributedModel.9.0.0.csproj" />
+    <ProjectReference Include="../../system.composition.hosting/9.0.0/System.Composition.Hosting.9.0.0.csproj" />
+    <ProjectReference Include="../../system.composition.runtime/9.0.0/System.Composition.Runtime.9.0.0.csproj" />
+  </ItemGroup>
+
+  <ItemGroup Condition="'$(TargetFramework)' == 'net9.0'">
+    <ProjectReference Include="../../system.composition.attributedmodel/9.0.0/System.Composition.AttributedModel.9.0.0.csproj" />
+    <ProjectReference Include="../../system.composition.hosting/9.0.0/System.Composition.Hosting.9.0.0.csproj" />
+    <ProjectReference Include="../../system.composition.runtime/9.0.0/System.Composition.Runtime.9.0.0.csproj" />
+  </ItemGroup>
+
+  <ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.0'">
+    <ProjectReference Include="../../system.composition.attributedmodel/9.0.0/System.Composition.AttributedModel.9.0.0.csproj" />
+    <ProjectReference Include="../../system.composition.hosting/9.0.0/System.Composition.Hosting.9.0.0.csproj" />
+    <ProjectReference Include="../../system.composition.runtime/9.0.0/System.Composition.Runtime.9.0.0.csproj" />
+  </ItemGroup>
+
+</Project>
