+++ src/referencePackages/src/system.composition.typedparts/9.0.0/system.composition.typedparts.nuspec
@@ -1,0 +1,33 @@
+<?xml version="1.0" encoding="utf-8"?>
+<package xmlns="http://schemas.microsoft.com/packaging/2013/05/nuspec.xsd">
+  <metadata>
+    <id>System.Composition.TypedParts</id>
+    <version>9.0.0</version>
+    <authors>Microsoft</authors>
+    <license type="expression">MIT</license>
+    <licenseUrl>https://licenses.nuget.org/MIT</licenseUrl>
+    <projectUrl>https://dot.net/</projectUrl>
+    <description>Provides container configuration and some extension methods for the Managed Extensibility Framework (MEF).</description>
+    <releaseNotes>https://go.microsoft.com/fwlink/?LinkID=799421</releaseNotes>
+    <copyright>© Microsoft Corporation. All rights reserved.</copyright>
+    <serviceable>true</serviceable>
+    <repository type="git" url="https://github.com/dotnet/runtime" commit="9d5a6a9aa463d6d10b0b0ba6d5982cc82f363dc3" />
+    <dependencies>
+      <group targetFramework="net8.0">
+        <dependency id="System.Composition.AttributedModel" version="9.0.0" exclude="Build,Analyzers" />
+        <dependency id="System.Composition.Hosting" version="9.0.0" exclude="Build,Analyzers" />
+        <dependency id="System.Composition.Runtime" version="9.0.0" exclude="Build,Analyzers" />
+      </group>
+      <group targetFramework="net9.0">
+        <dependency id="System.Composition.AttributedModel" version="9.0.0" exclude="Build,Analyzers" />
+        <dependency id="System.Composition.Hosting" version="9.0.0" exclude="Build,Analyzers" />
+        <dependency id="System.Composition.Runtime" version="9.0.0" exclude="Build,Analyzers" />
+      </group>
+      <group targetFramework=".NETStandard2.0">
+        <dependency id="System.Composition.AttributedModel" version="9.0.0" exclude="Build,Analyzers" />
+        <dependency id="System.Composition.Hosting" version="9.0.0" exclude="Build,Analyzers" />
+        <dependency id="System.Composition.Runtime" version="9.0.0" exclude="Build,Analyzers" />
+      </group>
+    </dependencies>
+  </metadata>
+</package>
