--- src/textOnlyPackages/src/system.composition/1.1.0/Customizations.props
@@ -1,17 +1,0 @@
-<Project>
-
-  <ItemGroup>
-    <!-- Fix project paths -->
-    <ProjectReference Remove="../../system.composition.attributedmodel/1.1.0/System.Composition.AttributedModel.1.1.0.csproj" />
-    <ProjectReference Remove="../../system.composition.convention/1.1.0/System.Composition.Convention.1.1.0.csproj" />
-    <ProjectReference Remove="../../system.composition.hosting/1.1.0/System.Composition.Hosting.1.1.0.csproj" />
-    <ProjectReference Remove="../../system.composition.runtime/1.1.0/System.Composition.Runtime.1.1.0.csproj" />
-    <ProjectReference Remove="../../system.composition.typedparts/1.1.0/System.Composition.TypedParts.1.1.0.csproj" />
-    <ProjectReference Include="../../../../referencePackages/src/system.composition.attributedmodel/1.1.0/System.Composition.AttributedModel.1.1.0.csproj" />
-    <ProjectReference Include="../../../../referencePackages/src/system.composition.convention/1.1.0/System.Composition.Convention.1.1.0.csproj" />
-    <ProjectReference Include="../../../../referencePackages/src/system.composition.hosting/1.1.0/System.Composition.Hosting.1.1.0.csproj" />
-    <ProjectReference Include="../../../../referencePackages/src/system.composition.runtime/1.1.0/System.Composition.Runtime.1.1.0.csproj" />
-    <ProjectReference Include="../../../../referencePackages/src/system.composition.typedparts/1.1.0/System.Composition.TypedParts.1.1.0.csproj" />
-  </ItemGroup>
-
-</Project>
