--- src/textOnlyPackages/src/system.composition/1.1.0/System.Composition.1.1.0.csproj
@@ -1,17 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>any</TargetFrameworks>
-    <PackageVersion>1.1.0</PackageVersion>
-    <AssemblyName>System.Composition</AssemblyName>
-  </PropertyGroup>
-
-  <ItemGroup Condition="'$(TargetFramework)' == 'any'">
-    <ProjectReference Include="../../system.composition.attributedmodel/1.1.0/System.Composition.AttributedModel.1.1.0.csproj" />
-    <ProjectReference Include="../../system.composition.convention/1.1.0/System.Composition.Convention.1.1.0.csproj" />
-    <ProjectReference Include="../../system.composition.hosting/1.1.0/System.Composition.Hosting.1.1.0.csproj" />
-    <ProjectReference Include="../../system.composition.runtime/1.1.0/System.Composition.Runtime.1.1.0.csproj" />
-    <ProjectReference Include="../../system.composition.typedparts/1.1.0/System.Composition.TypedParts.1.1.0.csproj" />
-  </ItemGroup>
-
-</Project>
