--- src/textOnlyPackages/src/system.composition/1.1.0/version.txt
@@ -1,1 +1,0 @@
-8321c729934c0f8be754953439b88e6e1c120c24
